Add smart links
Smart links, when a host is provided, add "_target='blank' and
rel='nofollow'" to all links that point to an external domain.

This assumes anything used as a url that does not parse as a URI will be
assumes to be a local link and is formatted as "/#{url}". If you do
provide a complete url, then it will convert anything on the same host
to be a relative link and will not add the new window fields.