Add initial block configuration
For now, this will work if you provide information to the block
configuration with a `render_value` method.

Eventually, it also needs to take a node/class/object instead that has
a `render` method.

It would also be nice if there were a "Block" class you could subclass
that would enforce the necessary methods. I am not sure if it would have
render and/or render_value. But it should be able to handle nested block
children and structurify them appropriately.