Move href into meta fields to simplify the API
As the API is slowly coming together, it's clear that the `href` is just
another type of meta field but on a link. By moving that field into the
meta fields, it removes the special rendering needed for links.

It did require updating all tests to include `href` as a meta field, but
that is fine.

I have debated whether or not the css_class should also be considered
another meta field, but I'm less sure about that one. The only place it
would matter is inside the `HtmlTag` class, which combines them
together, but it doesn't really simplify any of the APIs.