Add ability to render a block in complex ways
Initially, a node could easily be rendered with a node class with a
render method or a render_value lamdba.

However, it is possible to want to render a block that has children, in
which case creating a render method may prove difficult, because you
would lose the iterative/recursive aspect of rendering various blocks.

This provides the ability to add a "structure" key to a block config,
which can be composed of "value", "field", "block", and "blocks"
elements.

The 'value' element represents using a render value function to be used,
this can be used if you need to turn something like a "url" field into
an anchor link.

The 'field' element represents using the exact value on a block field to
render as a span.

The 'block' element represents rendering a single block using a block
config.

The 'blocks' element represents rendering child blocks of the same type
using a block config.

One thing that is not possible is changing the config for a given block
based on the page you are on. This can be accomplished with CSS.