Extract Item object to be used for inline items and blocks
Blocks and inline items function identically, taking a DatoCMS item of
some sort and rendering it based on a node, function, structure, etc.

The only difference is whether they are "inline" or not.

This commit extracts and item class which can be re-used for both and
builds blocks on top of it.

Also readme updates.